Bare bones tests
